textParsingPipeline/requirements.txt
fastapi
uvicorn
google-cloud-storage
python-docx
PyMuPDF
pandas
openpyxl
beautifulsoup4
requests